add: 댓글 component로 나누기
